score ???

+++ Assets/Scripts/ScoreManager.cs
@@ -7,7 +7,7 @@ public class ScoreManager : MonoBehaviour
 {
     [SerializeField] private Text text_score;
 
-    public int arret;
+    public float arret;
 
     public int total;
     // Start is called before the first frame update
